feat: make token distribution an option
